+++ CHANGELOG.md
@@ -42,6 +42,10 @@ Section Order:
 
 <!-- Your changes go here -->
 
+### Changed
+
+- Use AA's evelinks API for images
+
 ## [2.0.0] - 2026-02-26
 
 ### Changed